Main

Description

Módulo da Tarefa 2 que contém as funções relativas ao cálculo das coordenadas e orientação após a execução do primeiro comando no âmbito do Projeto de Laboratórios de Informática I.

Documentation

type Tabuleiro = [String]

type Pos = (Int, Int, Char)

outStr :: [String] -> String

main :: IO ()

tarefa :: [String] -> [String]

tarefa: Função que executa o primeiro comando num dado tabuleiro:

Caso o comando seja válido - O valor final são as coordenadas e a orientação depois de executado o comando.

Caso o comando não seja válido - O valor final é ERRO

selTab :: [String] -> [String]

selTab: Função que seleciona as linhas respetivas ao tabuleiro.

selPos :: [String] -> Pos

selPos: Função que seleciona a linha respetiva à posição inicial.

selCom :: [String] -> Char

selCom: Função que seleciona a linha respetiva aos comandos.

convertToString :: Pos -> String

convertToString: Função que converte a 'Posição' para uma String.

proximaPos :: Pos -> Pos

proximaPos: Função que dadas as coordenadas e a orientação retornam as coordenadas e a orientação após o deslocamento.

esquerda :: Pos -> Pos

esquerda: Dada a orientação e o comando E a função dá a orientação depois de executado o comando.

direita :: Pos -> Pos

direita: Dada a orientação e o comando D a função dá a orientação depois de executado o comando.

indice :: Pos -> Tabuleiro -> Char

indice: Encontra a letra do tabuleiro correspondente às coordenadas x e y.

altura :: Pos -> Tabuleiro -> Int

altura: Converte a letra de um tabuleiro (Char) na respetiva altura (Int). Exemplo: a=0 , A=0 , b=1 , B=1.

testacoord :: Tabuleiro -> Pos -> Bool

testacoord: Dado um tabuleiro e uma posição testa se é possível, ou seja, as coordenadas depois do comando estão dentro do tabuleiro.