Tarefa1_2019li1g126

Description

Este módulo define funções comuns da Tarefa 2 do trabalho prático.

Testes

testesT1 :: [(Int, Int, Int)]

Testes unitários da Tarefa 1.

| Cada teste é um tuplo (número de Pistas,comprimento de cada Pista do Mapa,semente de aleatoriedades).

Funções pré-definidas da Tarefa 1.

geraAleatorios :: Int -> Int -> [Int]

o geraAleatorios é uma funcao dada

gera :: Int -> Int -> Int -> Mapa

Funções principais da Tarefa 1. | gera npistas comprimento semente | n do geraALeatorios == 2*npistas*comprimentoPistas-2*npistas -- | a gera é a funcao principal

gerAux :: Int -> Mapa

funcao auxiliar para gera

retaTerra :: [Pista] -> [Pista]

a retaTerra é uma funcao que adiciona reta terra no inicio de cada pista

geraListaPistas :: Int -> [(Int, Int)] -> [[(Int, Int)]]

o geraListaPistas tem a funcao de gerar listas de pistas

agrupaPares :: [Int] -> [(Int, Int)]

recebe npistas o comprimento e a seed -> um mapa -> [PISTAS] -> [PECAS] --

escolhePiso :: Int -> Piso -> Piso

o primeiro numero do par e o tipo de piso, o segundo numero do par e o tipo da peca --

alturaFinal :: Peca -> Int

gera lista pisos gera , exatamente o que o nome diz, lista de pisos geraListaPisos :: [(Int,Int)] -> Piso -> [Piso] geraListaPisos ((piso,peca):t) pisoanterior = pisob: geraListaPisos t pisob where pisob = escolhePiso piso pisoanterior geraListaPisos [] pisoanterior = []

esta funcao recebe h que é a altura da pista neste momento e n que é a gama que determina o tipo de peca | como as rampas podem descer e/ou subir e como hmin=0 entao pus de parte estes casos | esta funcao calcula a altura final de uma peca

escolhePeca :: Int -> (Int, Int) -> Piso -> Peca

o escolhe peca calcula a peca dada atraves de dados anteriores

geraListaPecas :: Int -> Piso -> [(Int, Int)] -> [Peca]

a geraListaPecas, gera listas de pecas tal como o nome sugere