Tarefa1_2021li1g083

Description

Tarefa 1 do projeto de LI1 em 2021/22

Este módulo, referente à tarefa 1, contém definições Haskell para avaliar se um potencial mapa do jogo Block Dude é jogável de acordo com cinco regras. 1. Apenas uma peça por posição 2. Apenas uma porta no jogo 3. Não existem caixas flutuantes 4. O mapa tem que ter pelo menos um espaço vazio 5. Deve exisir chão ao longo do mapa

Documentation

validaPotencialMapa :: [(Peca, Coordenadas)] -> Bool

A função "validaPotencialMapa" verifica que as regras do jogo são cumpridas utilizando funções auxiliares individuais para cada regra

checkDuplos :: [Coordenadas] -> Bool

Regra 1

"checkDuplos" verifica, recursivamente, que numa lista de apenas coordenadas não existem pares ordenados repetidos de modo a satisfazer a regra número 1.

jCoordenadas :: [(Peca, Coordenadas)] -> [Coordenadas]

"jCoordenadas" é auxiliar da função "checkDuplos" e é responsável por obter apenas as coordenadas das peças descartando o seu tipo

contaPortas :: [(Peca, Coordenadas)] -> Int

Regra 2

A função "contaPortas" assegura que a regra número 2 é cumprida contando o número de portas que existem na lista do tipo (peca,coordenadas) atribuida na "validaPotencialMapa"

checkBase :: [(Peca, Coordenadas)] -> [(Peca, Coordenadas)] -> Bool

Regra 3

"checkBase" irá localizar as peças do tipo caixa no mapa e a sua auxiliar "checkBloco" irá verificar se a base dessa peça é outra caixa ou um bloco evitando assim a existência de caixas fluantes como manda a regra número 3

checkBloco :: Coordenadas -> [(Peca, Coordenadas)] -> Bool

"checkBloco" verifica se uma caixa tem uma base válida

jCoordenadasX :: [Coordenadas] -> [Int]

"jCoordenadasX" é auxiliar da função "indiceMatriz" e é responsável por seperar os pares ordenados da lista das coordenadas nume lista com apenas os elementos X

jCoordenadasY :: [Coordenadas] -> [Int]

"jCoordenadasY" é auxiliar da função "indiceMatriz" e é responsável por seperar os pares ordenados da lista das coordenadas nume lista com apenas os elementos Y

indiceMatriz :: [Coordenadas] -> Coordenadas

"indiceMatriz" obtém um par ordenado que corresponde ao indice da matriz criada pela "criaMatriz". Esta função identifica qual o maior elemento das listas criadas pelas funções "jCoordenadasX" e "jCoordenadasY" e retorna o indice da matriz

criaMatriz :: [(Peca, Coordenadas)] -> [Coordenadas]

Como já referido na defnição do checkVazios a função "criaMatriz" irá criar uma matriz de indice m*n e criar todas posições possiveis nesse mapa desde (0,0) até (m,n)

checkVazios :: [Coordenadas] -> [Coordenadas] -> Bool

Regra 4

"checkVazios" irá verificar se existem espaços vazios como pede a regra número 4. Esta função recebe a lista das coordenadas das peças no mapa dado (lista obtida pela "jCoordenadas") e uma matriz de indice m*n com todas as coordenadas existentes (inclusive as não listadas). A matriz criada pela "criaMatriz" serve para compararmos as coordenadas das peças do mapa com a mesma. Comparando a lista das peças do mapa inserido com a lista de todas as coordenadas possiveis iremos concluir que se sobrarem pares ordenadas na lista criada pela "criaMatriz" então existem espaços vazios.

primeiroChao :: [(Peca, Coordenadas)] -> (Peca, Coordenadas)

"primeiroChao" é responsável por detectar o primeiro bloco que serve de chão

chao :: [(Peca, Coordenadas)] -> (Peca, Coordenadas) -> Bool

Regra 5

"chao" verifica se há chão ao longo de todo o mapa, obedecendo à regra 5. Esta função considera que se os Blocos tiverem apenas uma ligação por esquina, terão que ter uma caixa a liga-los, como no exemplo do Q&A

delete :: Eq a => a -> [a] -> [a]

"delete" é uma auxiliar das funções do chao que elimina da lista a peça anteriormente comparada para a função não entrar em loop

chaoalt :: [(Peca, Coordenadas)] -> (Peca, Coordenadas) -> Bool

"chaoalt" é uma função alternativa à "chao" Esta função considera que os Blocos podem se ligar apenas pela sua esquina. Inativa de momento