Tarefa2_2019li1g126

Description

Este módulo define funções comuns da Tarefa 2 do trabalho prático.

Testes

testesT2 :: [(Int, Jogada, Estado)]

Testes unitários da Tarefa 2.

Cada teste é um triplo (identificador do Jogador,Jogada a efetuar,Estado anterior).

Funções principais da Tarefa 2.

jogada

Arguments

:: Int

-> Jogada

-> Estado

-> Estado
O Estado resultante após o jogador efetuar a jogada.

Efetua uma jogada.

jogada :: Int -- ^ O identificador do Jogador que efetua a jogada. -> Jogada -- ^ A Jogada a efetuar. -> Estado -- ^ O Estado anterior. -> Estado -- ^ O Estado resultante após o jogador efetuar a jogada.

trata-se da funcao principal, em que dividi em casos para facilitar o codigo e compreensao

jogadaDispara :: Int -> Jogador -> Estado -> Estado

esta funcao é a funcao jogada dispara, recebe um inteiro (o índice do jogador), o jogador (os dados do jogador), e o estado do mapa e da lista de jogadores, retornando esse estado ja atualizado | a funcao dispara analisa os dados do Jogador, nomeadamente se o numero de colas tem é maior que zero,para que possa disparar e se nao esta na primeira peca, pq nao pode colar um piso que nao existe- se sim entao dispara e perde uma municao, se nao nao dispara e retorna o estado do mapa

colaPiso :: Peca -> Peca

esta funcao é a que converte o piso para cola

identificaColaPecaAnterior :: Int -> Jogador -> Estado -> Peca

esta funcao é a identificaColaPeca Anterior,vai buscar a peca anterior quando um jogador dispara e converte o piso dessa peca para Cola

identificaPecaAnterior :: Int -> Jogador -> Estado -> Peca

esta funcao identifica a peca anterior (era necessaria para a funcao movimenta)

identificaPecaAtual :: Int -> Jogador -> Estado -> Peca

esta funcao identifica a peca atual (necessaria para o movimenta)

jogadaAcelera :: Int -> Jogador -> Estado -> Estado

esta é a jogadaAcelera, esta recebe um indice, ou seja, qual dos jogadores esta a efetuar esta jogada, depois recebe a informacao deste ,e o estado do mapa e dos jogadores, retornando o estado do mapa e da lista dos jogadores atualizado caso o jogador acelere

jogadaDesacelera :: Int -> Jogador -> Estado -> Estado

esta é a jogadaDesacelera, faz exatamente, em termos de raciocinio, o que a jogadaAcelera faz, so que as informacoes e os detalhes estao ajustados pois trata se do Desacelera

difalturas :: Peca -> Peca -> Jogador -> Double

l altura, m inclinacao e n é gravidade | para a esquerda +15graus | para a direita -15graus | a inclinacao maxima é 90 e a minima é -90 graus

a difalturas interpreta duas pecas dadas pelas funcoes identificaPecaAnterior e identificaPecaAtual para fazer a subtracao das alturas destas, que lhe é dada por uma outra funcao

alturaRampa :: Peca -> Jogador -> Double

a altura de retas é bastante intuitiva portanto nao foi necessario criar uma funcao auxiliar para tal, mas no caso das rampas torna se mais dificil, e portanto criei esta funcao,visualizando uma rampa como se fosse uma reta num grafico cartesiano, em que o b é a altura inicial, o a a diferenca entre a alturafinal e a altura inicial e o x, vem dos detalhes do jogador, que é a sua distancia

alturaPeca :: Peca -> Jogador -> Double

a altura peca serve para determinar a altura de uma determinada peca

inclinacaoPeca :: Peca -> Double

esta funcao serve para determinar a inclinacao de uma dada peca

jogadaMovimenta :: Direcao -> Int -> Jogador -> Estado -> Estado

esta funcao serve para analisar o comportamento e movimento de um jogador e adequadamente escolhe a consequencia para tal