Tarefa2_2021li1g083

Description

Tarefa 2 do projeto de LI1 em 2021/22

Este módulo, referente à tarefa 2, contém definições Haskell para a construção/desconstrução de um mapa válido já avaliado na Tarefa 1

Documentation

desconstroiMapa :: Mapa -> [(Peca, Coordenadas)]

Funções relativas à desconstrução do mapa

"desconstroiMapa" é a função responsável por transformar a grelha do mapa com todo o tipo de peças listadas no data Pecas numa lista do tipo (Peca, Coordenadas) sem o tipo Vazio. Esta função fá-lo através de duas auxiliares, "auxdesconstroiMapa" e "desconstroiLinha".

auxdesconstroiMapa :: Mapa -> Coordenadas -> [(Peca, Coordenadas)]

"auxdesconstroiMapa" invoca recursivamente a "desconstroiLinha" para ir desconstruindo uma coluna especifica da matriz, de seguida esta função volta a invocar a "desconstroiLinha" mas para trabalhar com uma coluna diferente.

desconstroiLinha :: [Peca] -> Coordenadas -> [(Peca, Coordenadas)]

A função "desconstroiLinha", no seguimento da "auxdesconstroiMapa", vai percorrer todas as linhas de uma determinada coluna e eliminar todas as peças do tipo Vazio e juntando todas as restantes com a devida posição à lista do tipo (Peca, Coordenadas). == Exemplo: >>>desconstroiLinha [[Vazio,Vazio,Bloco],[Porta,Vazio,Vazio],[Bloco,Bloco,Bloco]] [(Bloco,(0,2)),(Porta,(1,0)),(Bloco,(2,0)),(Bloco,(2,1)),(Bloco,(2,2))]

constroiMapa :: [(Peca, Coordenadas)] -> Mapa

Funções relativas à construção do mapa

A função "constroiMapa" , contrariamente à "desconstroiMapa", e recorrendo às suas auxiliares transforma uma lista do tipo (Peca, Coordenadas) numa grelha com todas as peças, o Vazio inclusivé.

auxConstroiMapa :: [(Peca, Coordenadas)] -> Int -> Coordenadas -> Mapa

"auxConstroiMapa" apenas articula as auxiliares de modo a obtermos o mapa devidamente construido em "constroiMapa"

compLinhaP :: [(Peca, Coordenadas)] -> [Peca]

"compLinhaP" é uma auxiliar da "auxConstroiMapa" que apenas remove as coordenadas à lista de pares (Peca, Coordenadas). == Exemplo: >>> compLinhaP [(Bloco,(0,2)),(Porta,(1,0))] [Bloco,Porta]

compLinha :: [(Peca, Coordenadas)] -> [Coordenadas] -> Coordenadas -> [(Peca, Coordenadas)]

"compLinha" compara uma lista dada pela divAndar (documentada abaixo) com todas as coordenadas possíveis no mapa. Nos casos em que ums peça não é declarada entre coordenadas, ou seja, quando a peça é do tipo Vazio, a "compLinha" insere um par (Vazio,(x,y)) no respetivo local.

Exemplo:

>>> compLinhaP [(Bloco,(0,2)),(Porta,(1,0)),(Bloco,(2,0)),(Bloco,(2,1)),(Bloco,(2,2))]
[[Vazio,Vazio,Bloco],[Porta,Vazio,Vazio],[Bloco,Bloco,Bloco]]

divAndar :: [(Peca, Coordenadas)] -> Coordenadas -> [(Peca, Coordenadas)]

A função "divAndar" procura os elementos com o mesmo y de coordenada e junta-os numa lista. == Exemplo: >>> divAndar [(Bloco,(0,2)),(Porta,(1,0))] [(Bloco,(0,2))]