Tarefa3_2019li1g126

Description

Este módulo define funções comuns da Tarefa 3 do trabalho prático.

Testes

testesT3 :: [Mapa]

Testes unitários da Tarefa 3.

Cada teste é um Mapa.

Funções principais da Tarefa 3.

desconstroi :: Mapa -> Instrucoes

Desconstrói um Mapa numa sequência de Instrucoes.

NB: Uma solução correcta deve retornar uma sequência de Instrucoes tal que, para qualquer mapa válido m, executar as instruções '(desconstroi m)' produza o mesmo mapa m.

NB: Uma boa solução deve representar o Mapa dado no mínimo número de Instrucoes, de acordo com a função tamanhoInstrucoes.

Gera Instrucao

gerarInstrucao :: Mapa -> Int -> Instrucao

dada uma lista verifica se a cabeca e igual a um tipo de peca e da a instrucao de acordo com a peca --

Gera Instrucoes

gerarInstrucoes :: Mapa -> (Int, Int) -> Int -> Instrucoes

mapa = [pista] = [[peca]] -- | gerarinstrucoes e uma funcao que tem que gerar uma [instrucao], ou seja instucoes, utilizando a funcao gerarInstrucao, listaSemPRimeiraInstrucao e a funcao bEs --

Retira Primeiro Elemento da Lista

listaSemPRimeiraInstrucao :: Mapa -> Mapa

dado o mapa, retira sempre a primeira instrucao para poder gerar a proxima instrucao --

Contador Pistas

bEs :: (Int, Int) -> Int -> Mapa -> (Int, Int)

um contador que serve para indicar a pista em que a peca se enconte, nomeadamente o [x] --

Gera as Instrucoes Finais

gInstrucoes :: Mapa -> Instrucoes

dado um mapa cria uma lista de instrucoes de acordo com a peca, usando a funcao gerarInstrucoes e a funcao dimensaoMatriz da Tarefa 0 --

Padroes Horizontais

padroesHorizontais :: Instrucoes -> Instrucoes

dada uma lista de instrucoes, vai verificar quais sao os padroes e comprimir assim, em um repete. Usa a funao nmrVezes para saber quantas vezes repete o elemento na

Contador Vezes de Elementos Repetidos

nmrVezes :: Eq a => [a] -> Int -> Int

contador do numero de vezes que uma determinada instrucao aparece repetido

Padroes Verticais

Padroes Verticais Desfasados