Tarefa3_2021li1g083

Description

Tarefa 3 do projeto de LI1 em 2021/22

Este módulo, referente à tarefa 3, contém definições Haskell para o output do mapa

Documentation

showMapa :: Mapa -> [String]

"showMapa" reúne os elementos do tipo Mapa para serem interpretados por "mostra"

desenha :: Peca -> String

A função "desenha" associa a cada um dos elementos do tipo Peca contidos na lista (Peca,Coordenadas) num caracter, colocando-os seguidos numa string para serem interpretados pelo "showMapa"

nconcat :: [String] -> String

"nconcat" separa as linhas da matriz correspondente ao mapa por um caracter "n" que representa a respetiva mudança no terminal.

auxaddJogador :: Jogador -> String

"auxaddJogador" atribui um caracter consoante a direção do jogador de modo a ser facilmente identificável visualmente no terminal

addJogador :: [String] -> Jogador -> Coordenadas -> [String]

A função "addJogador" percorre todas as colunas do mapa e coloca o jogador na respetiva posição, com a ajuda de "auxaddJogador" coloca-o virado para a direção correta.

auxJogador :: String -> Jogador -> Coordenadas -> String

"auxJogador" é auxiliar da "addJogador" e percorre todas as linhas do mapa e coloca o jogador na respetiva posição

Orphan instances

Show Jogo

Instance details

Methods

showsPrec :: Int -> Jogo -> ShowS

show :: Jogo -> String

showList :: [Jogo] -> ShowS