Tarefa4_2019li1g126

Description

Este módulo define funções comuns da Tarefa 4 do trabalho prático.

Testes

testesT4 :: [(Double, Mapa, Jogador)]

Testes unitários da Tarefa 4.

Cada teste é um par (tempo,Mapa,Jogador).

Funções principais da Tarefa 4.

passo :: Double -> Mapa -> Jogador -> Jogador

Avança o estado de um Jogador um passo em frente, durante um determinado período de tempo.

funcao predefinida por docentes

acelera :: Double -> Mapa -> Jogador -> Jogador

Altera a velocidade de um Jogador, durante um determinado período de tempo. acelera :: Double -- ^ O tempo decorrido. -> Mapa -- ^ O mapa utilizado. -> Jogador -- ^ O estado anterior do Jogador. -> Jogador -- ^ O estado do Jogador após acelerar.

move :: Double -> Mapa -> Jogador -> Jogador

Altera a posição de Jogador, durante um determinado período de tempo. move :: Double -- ^ O tempo decorrido. -> Mapa -- ^ O mapa utilizado. -> Jogador -- ^ O estado anterior do Jogador. -> Jogador -- ^ O estado do Jogador após se movimentar.

alturaFinal :: Peca -> Int

funcao que indica a altura final de uma peca

calculaSeInterseta :: Double -> Peca -> Mapa -> Jogador -> Bool

funcao que calcula se a trajetoria de um jogador vai intersetar com o chao num determinado tempo

sabeIntersecao :: Double -> Peca -> Mapa -> Jogador -> Ponto

esta funcao utiliza a funcao calculaIntersecao para saber se tal é verdade e depois, prossegue à determinacao dessas novas coordenadas

pecaresultante :: Double -> Jogador -> Bool

esta funcao decide se um jogador ultrapassa , com o tempo dado, se estara na peca seguinte

limitepeca :: Double -> Jogador -> Bool

esta funcao esta definida para identificar se no tempo dado, o jogador ultrapassa o limite peca

alturas :: Peca -> Jogador -> Double

esta funcao defne a altura de uma peca

velocidadeChao :: Jogador -> Double -> Mapa -> Jogador

A atualização da velocidade do jogador está dividida em dois casos: 1) quando o jogador está no chão e 2) quando o jogador está no ar. | No primeiro caso, a nova velocidade (v') do jogador é calculada através da seguinte fórmula:

velocidadeAr :: Jogador -> Double -> Jogador

No segundo caso (i.e., quando o jogador está no ar) além da nova velocidade, | é preciso atualizar a velocidade causada pela gravidade.

A nova velocidade do jogador (v') é calculada da seguinte maneira:

identificaproxima :: Jogador -> Mapa -> Peca

esta funcao identifica qual sera a seguinte peca na mesma pista

identifica :: Jogador -> Mapa -> Peca

esta funcao identifica a peca atual (necessaria para o movimenta), e identifica a peca em que este se movimenta de momento

alturaInicial :: Peca -> Int

funcao definida para definir a altura de uma peca

inclinacao :: Peca -> Double

funcao definida para identificar a inclinacao de uma dada peca

atrito :: Peca -> Double

esta funcao devolve os valores de atrito associado a cada piso