Tarefa4_2021li1g083

Description

Tarefa 4 do projeto de LI1 em 2021/22

Este módulo, referente à tarefa 4, contém definições Haskell relativas aos movimentos do Jogador

Documentation

correrMovimentos :: Jogo -> [Movimento] -> Jogo

A função "correrMovimentos" prepara uma série de movimentos para serem aplicados ao Jogador.

moveJogador :: Jogo -> Movimento -> Jogo

"moveJogador" é auxiliar de "correrMovimentos" e aplica um movimento especifico ao jogador. Esta função também identifica se o personagem carrega ou não uma Caixa.

checkPeca :: Mapa -> Coordenadas -> Coordenadas -> Peca

A função "checkPeca" identifica o tipo de peça presente numa coordenada especifica.

moveLeftFalse :: Mapa -> Jogador -> Jogo

Caso o jogador não carregue uma Caixa:

"moveLeftFalse" é a função responsável por movimentar o jogador para a esquerda quando este não carrega uma caixa.

moveRightFalse :: Mapa -> Jogador -> Jogo

"moveRightFalse" movimenta o jogador para a direita quando este não carrega uma caixa.

moveUpFalse :: Mapa -> Jogador -> Jogo

"moveUpFalse" permite ao jogador trepar blocos ou caixas.

interCaixaFalse :: Mapa -> Jogador -> Jogo

"interCaixaFalse" é responsável pela interação Jogador - Caixa, ou seja, caso este interaja com uma caixa, passará a carregá-la se possivel.

moveLeftTrue :: Mapa -> Jogador -> Jogo

Caso o jogador carregue uma Caixa:

"moveLeftTrue" é a função responsável por movimentar o jogador e a caixa que este carrega para a esquerda.

moveRightTrue :: Mapa -> Jogador -> Jogo

"moveRightTrue" movimenta o jogador e a caixa que este carrega para a direita.

moveUpTrue :: Mapa -> Jogador -> Jogo

"moveUpTrue" permite ao jogador trepar blocos ou caixas se estiver a carregar uma caixa.

interCaixaTrue :: Mapa -> Jogador -> Jogo

"interCaixaTrue" é responsável pela interação Jogador - Caixa, ou seja, como o jogador já carrega uma caixa, se este tentar apanhar outra, não irá conseguir.

moveCaixa :: Mapa -> Coordenadas -> Coordenadas -> Mapa

Auxiliares

"moveCaixa" é auxiliar das funções anteriores e permite executar os movimentos do jogador também na caixa para que ambos se movam corretamente

addremoveCaixa :: Mapa -> Coordenadas -> Coordenadas -> Mapa

A "addremoveCaixa" elimina a caixa de uma coordenada sempre que um movimento/interação é executado, de modo a não existir a mesma caixa em posições diferentes

auxaddremoveCaixa :: [Peca] -> Coordenadas -> Coordenadas -> [Peca]

"auxaddremoveCaixa" COMPLETAR!!!!!!!

gravidadeCaixa :: [(Peca, Coordenadas)] -> Mapa

"gravidadeCaixa" é responsável por colocar caixas flutuantes no chão sempre que o jogador as larga.

gravidadeCaixaTrue :: [(Peca, Coordenadas)] -> Mapa

COMPLETAR!!!!!!!

gravidadeJogador :: [(Peca, Coordenadas)] -> Jogador -> Jogador

"gravidadeJogador" permite ao jogador cair, ou seja, sempre que o jogador passe para um nivel mais baixo do que se encontra, a "gravidadeJogador" coloca-o no chão para que este não flutue no mapa

checkBase1 :: [(Peca, Coordenadas)] -> [(Peca, Coordenadas)] -> Coordenadas

"checkBase1" verifica se a base de uma caixa é outra caixa ou um bloco, pois esta não pode flutuar de acordo com a regra 3 da tarefa 1

findFloor :: [(Peca, Coordenadas)] -> Coordenadas -> Coordenadas

"findFloor" é uma auxiliar e permite à "gravidadeCaixa" achar onde se encontra o chão mais proximo para lá colocar o jogador.

findFloorTrue :: [(Peca, Coordenadas)] -> Coordenadas -> Coordenadas

"findFloorTrue" é uma auxiliar e permite à "gravidadeCaixaTrue" achar onde se encontra o chão mais proximo para lá colocar o jogador e a caixa que este carrega.