Tarefa6_2019li1g126

Description

Este módulo define funções comuns da Tarefa 6 do trabalho prático.

Funções principais da Tarefa 6.

bot :: Int -> Estado -> Maybe Jogada

Define um robot capaz de jogar autonomamente o jogo. bot :: Int -- ^ O identificador do Jogador associado ao robot. -> Estado -- ^ O Estado para o qual o robot deve tomar uma decisão. -> Maybe Jogada -- ^ Uma possível Jogada a efetuar pelo robot.

botAcelera :: Int -> Estado -> Maybe Jogada

esta função decide quando um bot deve acelerar, ou seja quando este tiver Chao False

identificaPiso :: Jogador -> Mapa -> Piso

esta funcao é capaz de identificar qual o piso onde o bot se encontra de momento

identificaPisoAbaixo :: Jogador -> Mapa -> Piso

esta funcao é capaz de identificar qual o piso onde o bot se encontraria se estivesse na pista diretamente abaixo

identificaPisoAcima :: Jogador -> Mapa -> Piso

esta funcao é capaz de identificar qual o piso onde o bot se encontraria se estivesse na pista imediatamente acima, se tal for possivel , é claro

segundocomponente :: Peca -> Piso

esta funcao pega numa peca e decifra qual o piso nesta

botPiso :: Int -> Estado -> Maybe Jogada

esta funcao é utilizada na funcao principal, é esta que avalia se o bot deve ou nao mudar de piso

avaliaPisos :: Jogador -> Mapa -> Int

esta funcao avalia quais dos pisos é o mais eficiente, usando as funcoes identifica e a atribuiPontos

atribuiPontos :: Piso -> Int

esta funcao atribui valores aos pisos, o piso que tiver melhor cotacao, ou seja, maior valor, vai ter sempre prioridade aos outros

avaliaAr :: Int -> Jogador -> Estado -> Maybe Jogada

esta funcao avalia se o angulo em que o jogador esta no ar esta dentro dos parametros, e reage de maneira adequada para que o jogador nao morra quando atingir o chao, vai ser utilizada na funcao principal

analisaJogadores :: Int -> Estado -> Maybe Jogada

esta funcao analisa todos os jogadores no mapa, e se estes tiverem a menos de 1.01 distancia do bot,na mesma pista, este dispara cola, esta funcao é utilizada na funcao principal

retiraPistaPiso :: Jogador -> (Int, Double)

esta funcao retira dos dados de um jogador para um par ordenado , respetivamente, os seguintes valores, a pista e a distancia a que esta esta nesta