Main

Description

Contém funções que detetam se o tabuleiro contém erros de sintaxe (ex: uma linha em branco, tabuleiro com digitos, ...). Caso ocorra algum erro de sintaxe é retornado o valor da linha onde ocorre o primeiro erro. Caso contrário é devolvida a String OK.

Documentation

main :: IO ()

main recebe o std input de um ficheiro.

Devolve a mensagem resultante da função process.

>>> "aaa\naaa\n0 0 S\nLASED"
OK

>>> "aaa\nbaa\n\n0 0 S\nLASED"
3

process :: [String] -> String

process executa todas as funções necessárias para a validação do tabuleiro, da posição inicial do robot e dos comandos que lhe serão aplicados. Executa cada uma delas sequencialmente, caso alguma falhe, a função process termina devolvendo uma mensagem contento o número da linha do ficheiro onde ocorreu o erro.

getBoard :: [String] -> ([Int], Int, [String], [String])

getBoard separa o tabuleiro das coordenas e dos comandos. Devolve também a largura de cada linha do tabuleiro, bem como a altura do tabuleiro.

>>> ["aaa","ba","1 0 S","LASDE"]
([3, 2], 2, ["aaa", "ba"], ["1 0 S", "LASDE"]

isBoard :: String -> Int -> Int

isBoard testa se uma linha pertence ou não ao tabuleiro (se todos os seus caráteres são letras). Caso algum caráter não seja uma letra é devolvido o valor 0. Caso contrario é devolvido o valor da largura dessa linha.

>>> "aaa"
3

>>> "aa2"
0   

isValidBoard :: Int -> Int -> [Int] -> Int

isValidBoard testa se o tabulero é valido (se todas as linhas têm a mesma largura). Caso todas as linhas tenham a mesma largura é devolvido o valor 0. Caso contrário é devolvido o número da linha onde a condição falhou. O primeiro argumento recebido corresponde à largura da primeira linha do tabuleiro.

>>> 3 1 [3, 3, 4]
3

>>> 3 1 [3, 3, 3]
0 

isVSintax :: String -> (Int, Int) -> Bool

isVSintax testa se as coordenadas são válidas. Para isso é necessário:

1 - que a String recebida contenha 2 espaços.

2 - que o tamanho da words (String recebida) seja 3.

3 - que a primeira e a segunda posições da words (String recebida) sejam inteiros sem o seguinte formato (ex: 00001) e que a terceira posição seja um caráter N, S, E ou O.

4 - que a primeira posição da words (String recebida) seja menor que a (largura do tabuleiro - 1) e que a segunda posição seja menor que a (altura do tabuleiro - 1).

>>> "1 0 S" (2, 2)
True

>>> "0001 0 S" (2, 2)
False

>>> "0 1 S" (2, 1)
False

countSpaces :: String -> Int

countSpaces conta o número de espaços de uma String.

>>> "1 0 S"
2

isVCoord :: String -> Int -> Bool

isVCoord testa se uma coordenada x ou y são válidas. Caso sejam inteiros e não estejam no seguinte formato (ex: 00001).

>>> "10"
True

>>> "0"
True

>>> "a"
False

>>> "01"
False 

iError :: String -> Bool

iError testa se os comandos se encontram na sintaxe válida, ou seja, se todos os caráteres da String recebida têm os seguintes valores: L, A, S, D ou E.

>>> "LAS"
True

>>> "As"
False