Main

Description

Conjunto de funções que calculam duas componentes do jogo: as posições de todos os blocos constituintes do tabuleiro e todas as posições que o robô ocupa enquanto executa as instruções. Tendo ambas as componentes, as funções deste módulo imprimem o código de uma página HTML, onde aparece o tabuleiro (blocos normais e lâmpadas) e o robô a movimentar-se segundo as instruções (andar, saltar, rodar para a esquerda ou para a direita e acender uma lâmpada).

Documentation

type LightsOff = (Int, Int)

LightsOff - coordenadas (x, y) de uma luz.

type Board = [String]

Board - caráteres que compõem o tabuleiro.

type Size = (Int, Int)

Size - largura e altura do tabuleiro.

type BoardInfo = (Board, Int, String, String, [LightsOff])

BoardInfo - informações do tabuleiro: Board, altura do tabuleiro, string com as coordenadas e orientação iniciais do robô, string com as instruções para o robô executar e lista com as coordenadas das lâmpadas por acender ([LightsOff]).

type Coord = (Int, Int, Orientation)

Coord - coordenadas do robô (segundo x e segundo y) e orientação.

type Position = (Int, Int, Float, Command)

Position - posição do robô com coordenadas x, y e orientação, antes da transformação para HTML.

type PositionP = (Board, Coord, String, Size, [LightsOff])

PositionP - parâmetros que influênciam o cálculo da próxima posição do robô: tabuleiro, coordenadas, linha de instruções, tamanho do tabuleiro e lista com as coordenadas das lâmpadas

type Bot = (Float, Float, Float, Float, Command)

Bot - posição do robô (coordenadas, orientação em radianos e comando a executar) após a transformação para o código HTML.

type Legs = (String, String)

Legs - código HTML que faz com que as pernas do robô se movimentem.

data Orientation

Orientation - orientações possíveis do robô.

Constructors

N

E

S

O

data Command

Command - comandos que o robô pode executar.

Constructors

W

J

JU

JD

LON

LOFF

IL

R

L

NOT

main :: IO ()

main recebe o std input de um ficheiro. Devolve uma string resultante da função process. Esta string devolvida é o código HTML que constrói o tabuleiro e os movimentos do robô.

process :: [String] -> String

process faz o processamento do tabuleiro, das coordenadas iniciais do robô, das instruções, das lâmpadas por acender e das dimensões do tabuleiro. Depois, faz os calculos necessários para determinar todas as posições que o robô percorre ao executar os comandos. No fim, imprime o código HTML que gera o tabuleiro e o robô.

separateBoard :: Board -> BoardInfo

separateBoard separa o conteúdo recebido, devolvendo as informações do tabuleiro.

>>> ["aaa", "Bbb", "ccC", "0 0 N", "ASDE"]
(["aaa", "bbb", "ccc"], 3, "0 0 N", "ASDE",[(2, 0), (0, 1)])

getLC :: String -> Int -> [Int]

getLC devolve as coordendas em x, das luzes presentes numa dada linha do tabuleiro.

>>> "aaAaB"
[2, 4]

getCoord :: [String] -> Int -> (Int, Int, Orientation)

getCoord devolve as coordenadas iniciais do robô em formato (Int, Int) e a orientação em formato Orientation.

toOrientation :: Char -> Orientation

toOrientation devolve a orientação do robô em formato Orientation.

height :: Int -> Float

height devolve a altura de um bloco (ou lâmpada) no formato para o código HTML.

convert :: Char -> Float

convert faz a transformação da altura do robô para o formato a ser impresso no código HTML.

getP :: PositionP -> [Position]

getP calcula todas as posições por onde o robô passa, a partir da sua posição inicial e das instruções.

getLightC :: (Int, Int) -> Size -> [LightsOff] -> Board -> (Command, [LightsOff])

getLightC verifica, para a posição onde o robô se encontra e para o comando L fornecido, se ele acende uma lâmpada, apaga uma lâmpada ou não é possível apagar ou acender uma lâmpada.

>>> (1, 1) (3, 3) [(0, 0), (1,1)] ["aaa", "aAa", "Aaa"]
LON

>>> (1, 1) (3, 3) [(0, 0)] ["aaa", "aaa", "Aaa"]
IL

getWJC :: Board -> Coord -> Size -> Char -> [Position]

getWJC devolve, para as instruções andar (A) e saltar (S), as coordenadas que o robô percorre.

>>> ["aaa", "bAa", "Aaa"] (0, 0, N) (3, 3) 'A'
[(0, 0, 0.25, W)]

>>> ["aaa", "bAa", "Aaa"] (0, 0, N) (3, 3) 'S'
[(0, 0, 0.5, JU), (0, 1, 0.5, JU)]

getNextP :: Orientation -> (Int, Int)

getNextP devolve, a partir da orientação do robô, o que se deve somar às coordenadas x e y quando a instrução é andar ou saltar. Isto para que calcule a posição seguinte.

isInvalid :: (Int, Int) -> Size -> Bool

isInvalid verifica se nas coordenadas atuais do robô é possível executar as instruções andar e saltar.

getWC :: Board -> (Int, Int) -> (Int, Int) -> Size -> [Position]

getWC devolve a lista com as posições que o robô percorre quando ele anda.

>>> ["aaa", "aAa", "aaa"] (0, 0) (0 , 1) (3, 3)
[(0, 1, 0.25, W)]

>>> ["aaa", "bAa", "baa"] (0, 1) (0 , -1) (3, 3)
[(0, 0, 0.5, W)]

getJC :: Board -> (Int, Int) -> (Int, Int) -> Size -> [Position]

getJC devolve a lista com as posições que o robô percorre quando ele salta.

>>> ["aaa", "aAa", "baa"] (0, 0) (0 , 1) (3, 3)
[(0, 1, 0.5, JD), (0, 1, 0.25, JD)]

>>> ["aaa", "bAa", "baa"] (0, 0) (0, 1) (3, 3)
[(0, 0, 0.75, J), (0, 0, 0.5, J)]

getXY :: Position -> (Int, Int)

getXY separa as coordenadas x e y da orientação e do comando a executar pelo robô.

locate :: Board -> (Int, Int) -> Int -> Char

locate localiza no tabuleiro, o caráter cuja posição corresponde à do robô.

>>> ["abc"] (1, 0) 1
'b'

rmL :: [LightsOff] -> (Int, Int) -> [LightsOff]

rmL remove da lista com as coordenadas de todas as lâmpadas, as coordenadas fornecidas.

>>> [(0, 0), (1, 1), (2, 3)] (1, 1)
[(0, 0), (2, 3)}

left :: Orientation -> Orientation

left é invocada quando a instrução a executar é E (Esquerda - left), alterando a orientação do robot.

right :: Orientation -> Orientation

right é invocada quando a instrução a executar é D (Direita - right), alterando a orientação do robot.

getRadians :: Orientation -> Float

getRadians devolve a orientação do robô em radianos.

convertL :: [Position] -> Float -> (Float, Float) -> Int -> [Bot]

convertL faz a conversão da lista com as posições do robô para uma lista com as posições prontas a serem impressas num código HTML.

convertLights :: [LightsOff] -> (Float, Float) -> Int -> [(Float, Float)]

convertLights faz a conversão da lista com as coordenadas inteiras das lâmpadas para uma lista com as coordenadas prontas a serem impressas num código HTML.

htmlHead :: String

htmlHead devolve o cabeçalho necessário para o código HTML.

htmlBodyShapes :: String

htmlBodyShapes devolve o código HTML que centra a janela com o tabuleiro e que contém as informações (dimensões e cores) dos blocos (que não são lâmpadas) do tabuleiro.

htmlBoard :: Board -> Float -> Float -> Float -> String

htmlBoard devolve o código HTML que contrói as linhas do tabuleiro.

htmlLines :: String -> Float -> Float -> Float -> Float -> String

htmlLines devolve o código HTML que constrói as colunas do tabuleiro. Invoca a função htmlLights quando o caráter que está a ler é uma maiúscula (lâmpada). Invoca a função htmlBlocks quando o caráter que está a ler é uma minúscula (bloco normal).

htmlBlocks :: Float -> Float -> Int -> Float -> Float -> String

htmlBlocks devolve o código HTML que constrói os blocos normais do tabuleiro.

htmlLights :: Float -> Float -> Int -> Float -> Float -> String

htmlLights devolve o código HTML que constrói as lâmpadas do tabuleiro.

auxRound :: Float -> Float

auxRound arredonda um número à milésima.

htmlBot :: String

htmlBot contrói o código HTML que constitui o robô.

htmlAnimatronics :: [Bot] -> String

htmlAnimatronics devolve todas as coordenadas x, y e z que o robô percorre, em forma de String.

htmlRotation :: [Bot] -> String

htmlRotation devolve todas as orientações do robô (segundo y), em forma de String.

htmlRoutePosition :: String

htmlRoutePosition devolve o código HTML que faz as routes das posições do robô.

htmlRouteOrientation :: String

htmlRouteOrientation devolve o código HTML que faz as routes das orientações do robô.

htmlColor :: [Bot] -> String

htmlColor devolve todas as cores que a antena do robô tem enquanto percorre o tabuleiro.

htmlEnding :: String

htmlEnding devolve o código final da página HTML.

htmlKey :: Float -> Float -> Int -> String

htmlKey devolve todas as Keys para os vários interpolators.

htmlLightsK :: [(Float, Float)] -> [Bot] -> String -> String

htmlLightsK devolve o colorInterpolator de cada lâmpada e as respetivas routes.

htmlLightsC :: (Float, Float) -> String -> [Bot] -> String

htmlLightsC devolve as cores de uma lâmpada, enquanto o robô percorre o tabuleiro.

auxLights :: (Float, Float) -> (Float, Float) -> Bool

auxLights compara duas coordenadas de lâmpadas e verifica se são iguais, com um desvio de 0.005.

convert2Walk :: [Bot] -> [Legs]

convert2Walk devolve uma lista com as orientações, em radianos, das pernas do robô.

htmlRouteOrientationLegs :: String -> String -> String

htmlRouteOrientationLegs devolve o código HTML que faz as routes das pernas do robô.

htmlRotationLegs :: Char -> [Legs] -> String

htmlRotationLegs devolve o código HTML para as orientações das pernas do robô.