Main

Description

Conjunto de funções que determinam a próxima posição do robot após a execução do primeiro comando. Começa-se por verificar se a instrução é possível de executar na posição actual. Se o comando for possível de executar, calcula-se a nova posição, apresentando-a no fim. Caso o comando não seja possível, é apresentada a mensagem ERRO.

Documentation

type Size = (Int, Int)

Size é o tipo de variável das dimensões da matriz (altura e largura).

type Coord = (Int, Int, Char)

Coord é o tipo de variável das coordenadas do robot.

type State = ([String], Coord, Char, Size)

State é o tipo de variável que contém o estado do robot.

main :: IO ()

main Recebe o std input de um ficheiro. Devolve a mensagem resultante da função process.

>>> "aaa\nbbb\nccc\n0 0 N\nASDE"
ERRO

>>> aaa\nbbb\nccc\n0 0 N\nSEAL"
0 1 N

process :: [String] -> String

process faz o processamento do tabuleiro e verifica se é possível executar o comando. Se for possível, apresenta as novas coordenadas. Caso contrário, apresenta a mensagem ERRO.

separateTab :: [String] -> ([String], Int, String, String)

separateTab separa o std input, devolvendo um tuplo com as linhas do tabuleiro, o número de linhas do tabuleiro, a linha com as coordenadas iniciais e a linha com os comandos.

>>> "aaa\nbbb\nccc\n0 0 N\nASDE"
(["aaa", "bbb", "ccc"], 3, "0 0 N", "ASDE")

getCoords :: [String] -> Coord

getCoords calcula as coordenadas iniciais. Recebe uma linha com as coordenadas inicias e devolve um tuplo com dois inteiros e um caráter.

>>> ["0","0","N"]
(0, 0, 'N')

locate :: [String] -> Coord -> Int -> Char

locate localiza um caráter no tabuleiro a partir das suas coordenadas, devolvendo-o no final.

>>> ["aaa", "bbb", "ccc"] (0, 0, N) 3
'c'

isValidCommand :: State -> Char -> Bool

isValidCommand chama a função necessária para verificar se o comando a excutar é válido na posição do robot.

>>> ["aaa", "bbb", "ccc"] (0, 0, N) 'S' 'c' (3, 3)
True

>>> ["aaa", "bbb", "ccc"] (0, 0, N) 'A' 'c' (3, 3)
False

walk :: State -> Bool

walk é chamada quando o comando a ser executado é A (Andar). Verifica se na posição actual a instrução é possível de executar. Devolve True caso seja possível.

jump :: State -> Bool

jump é chamada quando o comando a ser executado é S (Saltar). Verifica se na posição actual a instrução é possível de executar. Devolve True caso seja possível.

left :: Char -> Char

left é invocada quando a instrução a executar é E (Esquerda - left), alterando a orientação do robot.

right :: Char -> Char

right é invocada quando a instrução a executar é D (Direita - right), alterando a orientação do robot.

changePosition :: Coord -> Char -> String

changePosition é chamada quando a instrução é possível de executar. A partir do comando, poderá apresentar as coordenadas (quando não se alteram) ou chamar um função que as altere, apresentando-as depois.

>>> (0, 0, 'N') 'E'
(0, 0, 'O')

>>> (0, 0, 'N') 'S'
(0, 1, 'N')

>>> (0, 0, 'E') 'A'
(1, 0, 'E') 

moveRobot :: Coord -> Coord

moveRobot altera a coordenada x ou y, caso o comando S ou A sejam válidos. As novas coordenadas dependem da orientação do robot.

>>> (1, 0, 'O')
(0, 0, 'O')