Main

Description

Módulo da Tarefa 1 que contém as funções relativas à verificação dos tabuleiros no âmbito do Projeto de Laboratórios de Informática I.

Documentation

outStr :: [String] -> String

main :: IO ()

tarefa :: [String] -> [String]

tarefa: função que recebe a validade do tabuleiro « e que retorna a linha do erro ou Ok

selTab :: [String] -> [String]

Funções de seleção e teste do tabuleiro

selTab: função que filtra as linhas do tabuleiro

testaTab :: [String] -> Int

testaTab: função que testa a validade do tabuleiro recorrendo à função verificaTab e testarLetras

verificaTab :: Int -> Int -> [String] -> Int

verificaTab: função que esta se todas as linhas do tabuleiro têm o mesmo comprimento.

testarLetras :: String -> Bool

testarLetras: função que valida se o tabuleiro tem apenas letras na sua constituição

selPos :: [String] -> String

Funções seleção e teste das coordenadas da posição inicial

selPos: função que filtra a linha da posição inicial

testaPos :: String -> Int -> Int

testaPos: função que valida se a posição inicial tem 3 comandos, e que chama as restantes funções de validação.

validaCoord :: String -> Bool

validaCoord: função que verifica se a terceira coordenada é uma das letras

validaNum :: String -> Bool

validaNum: função que valida se as duas primeiras coordenadas são números

testax :: String -> String -> Int

Funções que validam se a posição

testax: valida se o comprimento de cada linha é menor do que o valor da primeira coordenada, ou seja, se a coordenada X está dentro do tabuleiro.

testay :: [String] -> String -> Int

testay: valida se o número total de linahs é menor do que o valor da segunda coordenada, ou seja, se a coordenada Y está dentro do tabuleiro.

selCom :: [String] -> String

Funções de seleção e teste dos comandos

selCom: seleciona a linha dos comandos

validaComV :: String -> Int

validaComV: verifica se existem apenas os comandos válidos: *A=Avançar *S=Saltar *L=Luz *E=Esquerda *D=Direita

valigaComV :: Num a => [t] -> a