Main

Description

Módulo da Tarefa 3 que contém as funções que executam todos os comandos fornecidos no tabuleiro e dão como output consoante o enunciado no âmbito do Projeto de Laboratórios de Informática I.

Documentation

type Tabuleiro = [String]

type Pos = (Int, Int, Char)

outStr :: [String] -> String

main :: IO ()

tarefa :: [String] -> [String]

stlampadas :: [String] -> Int -> [String]

stlampadas: Cria uma lista com as coordenadas de todas as lâmpadas do tabuleiro.

executa :: [String] -> String -> String -> [String] -> Int -> [String]

executa: Função recursiva que corre cada comando pela função tarefaB e que devolve o resultado final depois de exexutados todos os testes necessários à validação do input.

upLight :: String -> [String] -> [String]

upLight: Testa se uma determinada luz está acesa ou não:

Se está apagada -> Já pertence à lista

Se está acesa -> Ainda não pertende à lista

Caso esteja apagada a função remove as coordenadas dessa luz da lista. Caso esteja acesa a função adiciona as coordenadas dessa luz à lista. Se a lista está vazia -> Todas as luzes estão acessas

remove :: String -> [String] -> [String]

remove: Função auxiliar do upLight que encontra o elemento pretendido e remove-o da lista.

tarefaB :: [String] -> [String]

Funções relativas à Tarefa 2 já documentadas no respetivo ficheiro.

selTab :: [String] -> [String]

selPos :: [String] -> Pos

selCom :: [String] -> Char

convertToString :: Pos -> String

proximaPos :: Pos -> Pos

esquerda :: Pos -> Pos

direita :: Pos -> Pos

indice :: Pos -> Tabuleiro -> Char

altura :: Pos -> Tabuleiro -> Int

testacoord :: Tabuleiro -> Pos -> Bool