Main

Description

Módulo da Tarefa 4 que contém as funções que calculam os comandos necessários para completar um tabuleiro.

Documentation

outStr :: [String] -> String

main :: IO ()

type Pos = (Int, Int)

type Orient = Char

type Cmd = Char

type Tab = [String]

type Input = [String]

tarefa :: [String] -> [String]

tarefa: Função que dado um tabuleiro e a posição inicial de um robot calcula os comandos necessários para ligar todas as luzes do tabuleiro.

selTab :: [String] -> [String]

selTab: Função que seleciona o tabuleiro a partir da [String] inicial.

selPos :: [String] -> (Pos, Orient)

selTab: Função que seleciona a posição inicial a partir da [String] inicial.

stlampadas :: [String] -> Int -> [Pos]

stlampadas: Cria uma lista com as coordenadas de todas as lâmpadas do tabuleiro.

listaAlt :: Tab -> [[Int]]

listaAlt: Função que calcula todas as alturas de cada posição do tabuleiro.

nivel :: Char -> Int

nivel: Dado um char, retorna a altura correspondente. A altura de a é 0.

65 -> Corresponde ao código ASCII do caracter A.

97 -> Corresponde ao código ASCII do caracter a.

luzes :: [Pos] -> Pos -> Orient -> Tab -> [Cmd]

luzes: Função que a partir da lista das luzes vai sucessivamente acender todas as luzes de determinado tabuleiro.

[Pos] -> Corresponde à lista das posições das luzes do tabuleiro.

Pos -> Coresponde à posição inicial do robot.

move :: Pos -> Pos -> Orient -> Tab -> [Cmd] -> ([Cmd], Orient)

move: Função que move o robot dentro do tabuleiro. Em primeiro lugar acerta a coordenada X e em seguida a coordenada Y.

I -> No caso do robot não conseguir avançar retorna a I, que significa que é impossível.

testaAlt :: Int -> Int -> String

testaAlt: Função que testa se duas alturas forem iguais o robot devolve o comando A (Avançar), caso não sejam iguais devolve o comando S (Saltar). | Anteriormente já filtramos se a diferença entre as duas posições é maior do que 1 (caso impossível).

adiciona :: Orient -> Int -> Int -> [Cmd] -> Orient -> [Cmd]

adiciona: Função

roda :: Orient -> Orient -> [Cmd]

roda: Função que a partir da orientação do robot e da orientação pretendida para o robot vira o robot para a esquerda o número de vezes necessária até atingir a orientação pretendida.

esquerda :: Orient -> Orient

esquerda: Função que calcula a orientação do robot após rodar.

indice :: Pos -> Tab -> Char

indice2 :: Pos -> [[Int]] -> Int