Main

Description

Módulo da Tarefa 5 que contém as funções que possibilitam que seja gerado código html (com recurso ao X3DOM) de modo a apresentar a animação respetiva aos dados inseridos conforme indicado no eunciado no âmbito do Projeto de Laboratórios de Informática I.

Documentation

outStr :: [String] -> String

main :: IO ()

type Tab = [String]

type Pos = (Int, Int, Char)

type Input = [String]

type Comandos = String

type Orient = Char

tarefa :: Input -> [String]

tarefa: Função que junta todo o HTML estático com as funções que geram todo o código que varia de tabuleiro para tabuleiro.

selTab :: [String] -> [String]

selTab: Função que seleciona o tabuleiro a partir do input inicial.

selPos :: [String] -> Pos

selPos: Função que seleciona a posição inicial a partir do input inicial.

selCom :: [String] -> String

selCom: Função que seleciona os comandos a partir do input inicial.

pontoDeVista :: Tab -> String

pontoDeVista: Função que recebe as linhas correspondentes ao tabuleiro no input fornecido e gera uma linha com as coordenadas respetivas ao ponto de vista

nivel :: Char -> Int

nivel: Dado um char, retorna a altura correspondente. A altura de a é 0.

65 -> Corresponde ao código ASCII do caracter A.

97 -> Corresponde ao código ASCII do caracter a.

pontos :: Tab -> [String]

Funções que geram o tabuleiro

pontos: Função que juntamente com a função processa, a função processa2 e a função proLinha transforam a representação do tabuleiro.

["aabB",BcAa] -> ["00a","01a","02b","03B","10B","11c","12A","13a"].

processa :: Tab -> Tab

processa2 :: Int -> Tab -> [String]

procLinha :: Int -> Int -> String -> [String]

proc :: [String] -> [String]

proc: Função que gera o código HTML relativo ao tabuleiro.

coordenadas :: String -> String

coordenadas: Função que calcula as coordenadas e as coloca no formato correto do HTML.

tipoc :: String -> String

tipoc: Função que seleciona o tipo de cubo consoante esse ponto seja ou não um cubo com lâmpada.

fill :: String -> Int -> [String]

fill: Função que preenche o restante tabuleiro, todas as posições em baixo, apenas quando o nível é superior a 0.

invNivel :: Int -> Char

invNivel: Função que faz o contrário da função nivel. A função a partir de uma altura calcula a letra do tabuleiro correspondente.

movimento :: Input -> [String]

Funções que geram todo o movimento do Robot

movimento: Função que gera as linhas respetivas à animação do robot

posToOrient :: Pos -> Orient

posToOrient: Função que recebe um triplo do tipo Pos e retorna apenas o terceiro elemento (orientação)

radianos :: Orient -> Float

radianos: Recebe um Orient e retorna o valor em radianos que corresponde à orientação pretendida na animação

geraRadianos :: Float -> Comandos -> String

geraRadianos: Recebe um valor em radianos (respetivo à orientação da posição inicial) e uma string de comandos e gera uma string com as keyframes da rotação do robot ao longo da animação consoante os comandos

converte :: Tab -> Pos -> String

converte: Função usada para gerar as key frames respetivas à translação do robot na animação

calculat :: Float -> Float -> Float -> [String]

calculat: Cria uma lista com os tempos respetivos às key frames da translação do robot na animação

posicao :: Tab -> Pos -> Comandos -> [Pos]

posicao: Cria uma lista com as posições respetivas às key frames da translação do robot na animação

verifica :: Tab -> Pos -> Bool

verifica: Testa se uma posição está dentro do tabuleiro ou não

move :: Tab -> Pos -> Pos

move: Dada uma posição, retorna a posição resultante da execução do comando S (saltar) ou A (avançar)

esquerda :: Pos -> Pos

esquerda: Recebe uma posição e retorna uma posição com as mesmas coordenadas, mas com a orientação à esquerda da anterior

direita :: Pos -> Pos

direita: Recebe uma posição e retorna uma posição com as mesmas coordenadas, mas com a orientação à direita da anterior

altura :: Tab -> Pos -> Int

altura: Calcula a altura do tabuleiro em determinada posição

selectx :: Pos -> String -> Char

selectx: Recebe uma posição e uma string (linha do tabuleiro) e retorna o caractere correspondente à coordenada x da posição fornecida

selecty :: Pos -> Tab -> [Char]

selecty: Recebe uma posição e as strings respetivas ao tabuleiro e retorna a linha correspondente à coordenada y da posição fornecida

robot :: [[Char]]

tb1 :: [[Char]]

tb2 :: [[Char]]