Main

Documentation

data Ui

menus

Constructors

Menu Int

data Npistas

numero de pistas

Constructors

Npista Int

data Cpistas

comprimento de pistas

Constructors

Cpista Int

type EstadoGloss = (Ui, Npistas, Cpistas, Estado, [Picture])

estado que conjuga tudo

mapa0 :: Mapa

mapa1 :: Mapa

mapa2 :: Mapa

mapa3 :: Mapa

mapa4 :: Mapa

colorir :: Piso -> Int -> Color

Jogo = 2

Serve para dar a cor a's Pecas

colorirParede :: Piso -> Int -> Color

Serve para dar a cor a's Paredes quando a altura da peca e' superior a 1

drawPeca :: Peca -> Float -> Int -> Picture

Serve Para desenhar uma Peca

drawRecta :: Int -> Piso -> Int -> Picture

Desenha a peca que tem como construtor Recta

paredeReta :: Int -> Piso -> Int -> [Picture]

Desenha a parede das Rectas

rampaDescer :: Picture -> Picture

Troca as coordenadas da Rampa, espelhando assim a Rampa que tinha sido desenhada

drawRampa :: Int -> Int -> Picture

Desenha a peca que tem como construtor Rampa

paredeRampaDescer :: [Picture] -> [Picture]

Desenha a parede das rampas a descer

paredeRampa :: Int -> Int -> Piso -> Int -> [Picture]

Desenha a parede das Rectas

drawPista :: Pista -> Float -> Int -> [Picture]

Desenha uma Pista com o auxilio da funcao drawPeca

drawMapa :: Mapa -> Int -> [Picture]

Desenha um mapa com a ajuda recursiva da funcao drawPeca

transformaPic :: [Picture] -> Picture

Transforma uma [Picture] gerada pelo drawMapa no Tipo Picture

colocarMapa :: [Picture] -> Picture

colocacao do mapa com coordenadas

estadoInicialGloss :: [Picture] -> EstadoGloss

estado gloss, ajuda para criar menu

desenhaEstadoGloss :: EstadoGloss -> Picture

funcao que desenha o estado gloss

reageTeclasJogo :: Event -> Estado -> Estado

funcao que aplica funcoes as teclas e faz com que estas tenham impacto no jogo

reageEventoGloss :: Event -> EstadoGloss -> EstadoGloss

reageEventoGloss serve para atualizar o estadoGloss, ou seja o menu,caso de evento, ou seja, o jogador utilizar as teclas

drawPlayers :: EstadoGloss -> [Picture]

funcao que desenha jogadores

jogadorA :: EstadoGloss -> Picture

um jogador

alturaDoJogador :: Int -> Int -> Jogador -> Double

diz qual e a altura do jogador naquela peca ....usar a distancia do limite da peca com as alturas inicial e final pra descobrir a altura atual

drawPlayer :: Jogador -> Peca -> EstadoGloss -> Picture

Desenha um jogador

jogador1 :: Jogador

jogador1 = (Jogador 0 0 0 0 (Chao True))

distanciaLimPeca :: Jogador -> Double

funcao que calcula a distancia de um jogador ao limite de uma peca

reageAoTempo :: Float -> Estado -> Estado

funcao que faz com que o jogo reaja à passagem do tempo

reageAoTempoGloss :: Float -> EstadoGloss -> EstadoGloss

esta funcão tem a mesma funcionalidade que a reageAoTempo mas atualiza o estadoGloss

fr :: Int

framerates

backgroundColor :: Color

cor de fundo

main :: IO ()

Função principal da Tarefa 5.

NB: Esta Tarefa é completamente livre. Deve utilizar a biblioteca gloss para animar o jogo, e reutilizar de forma completa as funções das tarefas anteriores.